Organizing the create tour services calls.  Not fully tested.
